dom/tests/mochitest/geolocation/test_clearWatch.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=478911
-->
<head>
  <title>Test for watchPosition and clearWatch</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=478911">Mozilla Bug 478911</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

var watchID;
var hasBeenCleared = false;
function failureCallback(error)
{
  ok(0, "we should not be seeing failures from this watchPosition");
}

function successCallback(position) {
  if (hasBeenCleared == true) {
    // we should not be called here
    ok(0, "successCallback was called after clear");
  }
}

function clearWatch() {
  navigator.geolocation.clearWatch(watchID);
  hasBeenCleared = true;
}

function testAccepted() {
  SimpleTest.finish();
}


/** Test for Bug  **/

SimpleTest.waitForExplicitFinish();

watchID = navigator.geolocation.watchPosition(successCallback, failureCallback, null);

setTimeout(clickNotificationButton, 10, kAcceptButton);

setTimeout(clearWatch, 50);

// wait for position changes
setTimeout(testAccepted, 1000);

</script>
</pre>
</body>
</html>
